platforms: nanopi: Very preliminary version of nanopi_adaptor.go
Remove chip revision check which was used for RPi (doesn't apply here).
Replace with taking the chip revision as a parameter to the Adaptor
intializer. This must be checked against a constant list of supported
adaptors (currently only NEOv1.2). This means that the initializer must
also return an error type in case an unsupported revision string is
supplied.

This is untested and misses support for perihperals such as UART, I2S
etc.